[core-amqp] request response link should reject when sender error (#23646)
When the remote closes the sender with an error we currently just log the error logged as a warning and are not doing anything else. Since we won't receive any response, the send operation eventually times out, rejecting with `OperationTimeoutError`. As `OperationTimeoutError` is retryable, we would get the same time out error in each retry again, even though some sender errors are not retryable, for example, `MessageSizeExceeded` error in issue #23623.

This PR rejects the promise of request on sender_error events so that it fails with proper errors. These errors still go through the retry logic, but should be failing faster.

### Packages impacted by this PR
`@azure/core-amqp`