[Tables] Fix problem when concurrent transactions are submitted (#24190)
### Packages impacted by this PR
@azure/data-tables

### Issues associated with this PR
https://github.com/Azure/azure-sdk-for-js/issues/22341

### Describe the problem that is addressed by this PR
If `submitTransaction` is called concurrently, concurrent transactions step into each other's feet, causing one of the transactions to win over the others and be sent multiple times.

### What are the possible designs available to address the problem? If there is more than one possible design, why was the one in this PR chosen?

Currently, we have a cached TransactionClient which we use internally to translate the transaction actions into the batch request body. This causes concurrent transactions to use the same resource causing issues. The fix in this PR removed the cached transaction client and creates a new one per transaction to ensure isolation between transactions

### Are there test cases added in this PR? _(If not, why?)_
YES