[core-lro] Better detection for operation failure (#24566)
### Packages impacted by this PR
@azure/core-lro

### Issues associated with this PR
N/A

### Describe the problem that is addressed by this PR
The LRO is currently considered failed when any arbitrary error is raised during polling. However, the operation shouldn't be considered failed unless the raised error is a RestError. For example, an `AbortError` raised by a cancellation token doesn't imply that the LRO has failed.

### What are the possible designs available to address the problem? If there are more than one possible design, why was the one in this PR chosen?
I did a simple check for `error.name === RestError`. Alternatively, we could import the `isRestError` helper from `@azure/core-rest-pipeline` but there is no need to pull in `@azure/core-rest-pipeline`.

### Are there test cases added in this PR? _(If not, why?)_
No but updated existing ones.

### Provide a list of related PRs _(if any)_
N/A

### Command used to generate this PR:**_(Applicable only to SDK release request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so, create an Issue in the [Autorest/typescript](https://github.com/Azure/autorest.typescript) repository and link it here)_
- [x] Added a changelog (if necessary)