[Event Hubs] Rewrite partition receiver (#24731)
# Re-implementing the Event Receiver
This PR re-implements the event receiver using promises and a single
queue to fix an ordering issue and to correct waiting behavior.
## Problem Statement [Issue #23993]
A customer reported that the list of events passed into the
`processEvents` callback is not always ordered by `sequenceNumber`. This
leads to processing the events in a wrong order. The customer provided a
sample that prints an out of order message when the `sequenceNumber` of
received messages is not in order and I confirm that I see the message
printed sometimes.
## Analysis
The customer-provided callback, `processEvents`, gets called every time
a batch of events is received from the service. This batch is coming
from a single partition. Events are ordered within a partition by their
`sequenceNumber`, and events received by `processEvents` should be in
the same order. However currently, the list of events the
`processEvents` callback gets called on is not always in-order. Upon
further investigation, it was found that the library implements a
complex logic to read events from the service. It maintains two queues
for reading events, one for building a batch of events that will be sent
to the next call of the `processEvents` callback, and another for when
errors occur or there are no active listeners. The coordination to read
events from the two queues is subtle and is the source of the ordering
bug.
## Re-design
The most straightforward way to simplify this design and to ensure
ordering is to use a single queue and add incoming events to it in the
order they're received. Reading from this queue is as simple as the
following:
- If the queue contains any events, check if their count is already the
`maxMessageCount` or more:
  - If yes, remove `maxMessageCount` events and return them immediately
- If no, wait for a few milliseconds and then remove up to
`maxMessageCount` and return them
- If the queue doesn't contain any events, wait until the
`maxWaitTimeInSeconds` and then return an empty list, or until one or
more event arrive and then return those

### Abstraction

The idea is concisely captured by `waitForEvents`, a newly introduced
function that races a list of promises, one for each of the scenarios
listed above:
https://github.com/Azure/azure-sdk-for-js/blob/10826927554e7254dce0a4849f1e0c8219373522/sdk/eventhub/event-hubs/src/eventHubReceiver.ts#L733-L739

The first promise resolves right away and is returned if the queue
already has `maxMessageCount` events or more. It corresponds to the
first scenario listed above.

The second promise is created by the `checkOnInterval` function. The
promise is resolved only if the queue has any events in it. Otherwise,
it keeps checking every number of milliseconds. Note that chained to it
is a timer promise that waits another number of milliseconds to give the
service a chance to send more events. This corresponds to the second
scenario listed above.

The third promise is a simple timer promise that is resolved after the
`maxWaitTime` has elapsed. This promise corresponds to the third
scenario.

### Rewrite

In addition to some other minor improvements, the `receiveBatch` method
is concisely rewritten using that abstraction as follows:
https://github.com/Azure/azure-sdk-for-js/blob/10826927554e7254dce0a4849f1e0c8219373522/sdk/eventhub/event-hubs/src/eventHubReceiver.ts#L578-L628

Notice that the chain of promises makes the algorithm simple to read: a
link is established first, credits are added to it as needed, and then
the waiting starts.

Also, notice that at this point, no actual events were read from the
queue yet, all what this does is waiting until one of the promises
resolve. The actual reading from the queue is thened to that chain so
that it happens only after everything else is said and done. For
example, if an error occurred, it should be handled and we don't want to
prematurely mutate the queue. The reading from the queue is as simple as
the following:
https://github.com/Azure/azure-sdk-for-js/blob/10826927554e7254dce0a4849f1e0c8219373522/sdk/eventhub/event-hubs/src/eventHubReceiver.ts#L630

## Other changes

### Exporting `core-util`'s `createAbortablePromise`

This function was added in
https://github.com/Azure/azure-sdk-for-js/pull/24821 and proved to be
useful in this re-write so I am exporting it. I am planning on using it
in core-lro too.

### Updating tests

There are two tests updated, one for authentication and one for
returning events in the presence of retryable and non-retryable errors.

In the former, the receiver is expected to receive events after the auth
token has been invalidated but not yet refreshed. However, I am
observing that a disconnected event has been received at that moment and
the receiver has been deleted. The old receiver's behavior is to
continue receiving despite the deletion but the new one's behavior
correctly cleans up the receiver. I deleted this expectation for now.

In the latter, the test forces an error on the receiver after 50
milliseconds but the receiver already finishes around 40 milliseconds,
so I updated the forced error to happen sooner, at 10 milliseconds:
https://github.com/Azure/azure-sdk-for-js/blob/10826927554e7254dce0a4849f1e0c8219373522/sdk/eventhub/event-hubs/test/internal/receiveBatch.spec.ts#L107

Finally, a couple test suites were added for `waitForEvents` and
`checkOnInterval` functions.

## Updates in action

Live tests succeed
[[here](https://dev.azure.com/azure-sdk/internal/_build/results?buildId=2201768&view=results)].
Please ignore the timeout in the deployed resources script in canary, it
is an unrelated service issue, see
[[here](https://dev.azure.com/azure-sdk/internal/_build/results?buildId=2198994&view=results)].

A log for how the updated receiver behaves when used by the customer
sample can be found in
[log2.txt](https://github.com/Azure/azure-sdk-for-js/files/10775378/log2.txt).
Notice that the out of order message was never printed.

## Reviewing tips

The changes in `eventHubReceiver.ts` are too many and the diff is not
easily readable. I highly suggest to review
https://github.com/Azure/azure-sdk-for-js/pull/24731/commits/10826927554e7254dce0a4849f1e0c8219373522
instead because it is on top of a deleting commit so there is no diff to
wrestle with. The main changes are in `receiveBatch` but please feel
free to review the rest of the module too.