[Identity] Refining the Workload identity integration logic for ManagedIdentity Credential (#25010)
### Packages impacted by this PR
@azure/identity

### Issues associated with this PR
None. Related PR - https://github.com/Azure/azure-sdk-for-js/pull/24830

### Describe the problem that is addressed by this PR
- This PR makes the workload identity credential skip the entire
token-caching logic in in-house implemwntation of Managed Identity since
it's already taken care of by the Client Assertion Credential. Doing
this prevents the tenant id and authority host from being over-written
by the in-house implementation of ManagedIdentity and msal integration
from the ones user provided.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)