[Identity] Workload identity through DefaultAzureCredential fix for reading env vars (#25024)
### Packages impacted by this PR
@azure/identity

### Issues associated with this PR
Related - https://github.com/Azure/azure-sdk-for-js/issues/23326

### Describe the problem that is addressed by this PR
DefaultAzureCredential when passed the three env vars for client id,
tenant id and federated token file path was not passing the env vars to
WorkloadIdentityCredential and was just passing the
CredentialUnavailableError when the required parameters weren't present.
There were multiple challenges that I came across while working on a
design for this -
1. The CredentialUnavailableError or any other error thrown from the
WorkloadIdentity Credential - should immediately try the next credential
in the chain and not just exit after throwing.
2. The WorkloadIdentity should accept the env vars when passed in
through the DefaultAzureCredential client.
3. The constructor for WorkloadIdentity should match with the
constructor for DefaultAzureCredential. This was tricky to resolve as
well.
4. Next, the CredentialUnavailableError should be thrown from the
getToken() call and not from within the constructor, otherwise the next
credential in the chain will not run.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?

- Ideally the only constructor design for `WorkloadIdentityCredential`
should be to have the tenantId, clientId and federatedTokenFile path to
be passed through as required parameters.
- But the challenge for this constructor being accepted as a
"DefaultAzureCredentialConstructor" remains. Due to this I created an
overload constructor with just the options bags as
"WorkloadIdentityDefaultCredentialOptions" and made this options bag as
optional in the overload constructor. I made this constructor as
internal and hidden, since we want to dissuade customers from directly
using this with the `WorkloadIdentityCredential()`.
- Next, for the original constructor, I made the required params as
required but in the options bag. Finally created an implementation
constructor with unions of these 2 types.

- ### Are there test cases added in this PR? _(If not, why?)_


### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [x] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [x] Added a changelog (if necessary)

---------

Co-authored-by: Matthew Podwysocki <matthewp@microsoft.com>