[Communication] PhoneNumbers - Simplify status code assertions. (#25194)
Due to a small backend change, tests that validate operations on
unauthorized or invalid phone numbers now return status code 403, as
opposed to 400 or 404 previously. To simplify the code of the test, the
assertion now checks that the request has been rejected due to a client
error, which is validated by checking if the status code falls under the
4xx bucket.

While it should be enough to validate that the operation itself fails,
the check for client error was added to ensure that the request is
indeed being rejected by the server instead of it failing due to another
unrelated reason, which would usually be represented by a 5XX error.