[ServiceBus] Fix a race condition in generating replyTo and initializing link (#25279)
When a link is not ready for sending management request we will
initialize it
and a unique `replyTo` address is generated. The problem is the two
steps are
separate and only the link initialization is guarded by a lock to ensure
the
link is only initialized once. When multiple management requests are
made at the
same time, multiple `replyTo` addresses could be generated but only one
is
associated with the link, resulting in errors like

```
ServiceBusError: The response link '0f6cd6e7-16d1-034a-b661-5d632917cd50' was not found in the entity management node $management.
```

This PR groups the two parts into one locked operation, thus preventing
multiple
`replyTo` addresses from being generated for a new link.

- add a failing test without fix
- Move initialization code out of `_makeManagementRequest()` into a
locked block together with code to generate unique `replyTo`