[ServiceBus] add timeout around receiver drain when receiving messages (#25355)
This is a port of PR https://github.com/Azure/azure-sdk-for-js/pull/8239

From original PR description:
>This change attempts to remove an edge case where `receiveMessages` may
hang indefinitely.
>
>Currently, once the `BatchingReceiver` (`receiveMessages`) reaches
either the `maxMessageCount` or `maxWaitTime`, it will drain any
remaining credits on the receiver link and then return the messages.
>
>However if for some reason the service was unable to report back that
credits were drained, the messages would never be returned and the
operation would appear to hang indefinitely.
>
>Now the receiver will wait a maximum of 200 ms for the service to
indicate that credits are drained. If this timeout is reached, then the
received messages are returned from `receiveMessages`.
>
>I chose 200 ms because in my local testing, it typically took ~80 ms
between when the receiver triggered a drain and when the
`receiver_drained` event was fired. This gives an additional 150%
buffer. I had considered using `newMessageWaitTimeoutInSeconds` instead,
which currently is set to 1 second, but was trying to minimize how long
we'd wait so that message locks wouldn't need to be renewed.

In addition to porting the changes, I also update some existing tests to
work well with the new behavior, and reduced their execution time by
shorten the max waiting time.