[storage-blob] Fix paths in client instead of using a pipeline policy (#25381)
### Packages impacted by this PR

`@azure/storage-blob`

### Describe the problem that is addressed by this PR

When generating the client library, the OpenAPI input specification uses
path parameters to model sending the container and blob ids to the
service. However, since these ids require special handling to be escaped
properly, the SDK manages the complete URL itself.

This causes the path parameters to be duplicated when processed by
core-client. To avoid this problem, I previously used a pipeline policy
that removed the duplicate information.

While migrating `storage-file-datalake`, I thought of a different
solution, namely that of having the client itself modify the
OperationSpec directly before the request is generated. This results in
less work at runtime and is a bit cleaner in terms of making each
operation request.

This PR ports the same approach to storage-blob.