[ai-form-recognizer] Support API version 2023-02-28-preview (#25480)
WIP

This PR updates @azure/ai-form-recognizer to support API version
`2023-02-28-preview`. It updates the SDK version to 4.1.0-beta.1.

It also removes the test recordings which have been onboarded to the
asset sync tool.

The changes are described in CHANGELOG.md. All generated files (prebuilt
models) have been regenerated based on the latest model schemas. The
vast majority of changes are deleted recordings or regenerated models.
See the `api.md` file for changes to the API.