[design/dev-tool] automated migration framework (#25044)
This pull request adds a design document for an automated migration
framework that, if accepted, will be integrated into `dev-tool`.

I've come to view a tool like this as required to support continued
incremental updates across the monorepo.

I won't describe the tool too much in this description, since it's
described in the included Markdown file, but it basically works like a
git merge:

- dev-tool tracks a list of migrations, `dev-tool migrate --list` will
show the ones that haven't been applied to a package.
- you "merge" the migrations into a package using `dev-tool migrate`.
Migrations can be fully or partially automatic _OR_ manual. The tool
will tell you what to do.
- if the migration fails, it is suspended and you have the chance to fix
it and then run `dev-tool migrate --continue` and it will re-check it,
or you can run `dev-tool migrate --abort` to cancel the process.
- if a migration succeeds, a commit is created for it in the current
branch and the next migration is run. Migrations are run in order
starting from the oldest.

This PR is a work in progress pending comments.