Added Cosmos Diagnostics to client objects. (#25603)
### Packages impacted by this PR
@azure/cosmo

### Issues associated with this PR
[#21177](https://github.com/Azure/azure-sdk-for-js/issues/21177)

### Describe the problem that is addressed by this PR
These changes add `CosmosDiagnostics` to `@azure/cosmos` client objects.
This is second PR for diagnostics feature. Next upcoming changes are -
Query and batch. Here I have marked `diagnostics` field as optional in
`ResourceResponse` and `Response` objects, to make PR compliable, these
fields will be marked necessary in upcoming PRs.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?


### Are there test cases added in this PR? _(If not, why?)_
No, in the interest of keeping the PR manageable, Testcases will be
merged in a separate PR.

### Provide a list of related PRs _(if any)_


### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_

### Checklists
- [ ] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [ ] Added a changelog (if necessary)

---------

Co-authored-by: Vikas Singh <vikassingh@microsoft.com>