[core] add proper parameter type for TrustedTypePolicy `createHTML` callback (#25813)
### Packages impacted by this PR
`@azure/core-xml`, `@azure/core-http`

### Issues associated with this PR
Fixes issue #24872.

### Describe the problem that is addressed by this PR

Upgrading to @trusted-types v2.0.3 caused compilation error. While it is
breaking, it surfaces an issue that we can address by adding proper
parameter type.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
Pinning to v2.0.2 which isn't desirable as we want to stay on latest
dependency versions.