[@azure/eventgrid] Adding new events for EG Version 4.11.1 (#25834)
### Packages impacted by this PR
@azure/eventgrid

### Issues associated with this PR
NA

### Describe the problem that is addressed by this PR
The EventGrid Service Team has added 20 new events. The SDK must be
regenerated to include the code changes related to these 20 new events.
The SDK version must be updated and a out-of-band release has been
planned for this service.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
There are no specific/complex design scenarios for this task. It is a
straightforward regenerate and some standard changes to the custom layer
of the code.

### Are there test cases added in this PR? _(If not, why?)_
No. This item is standard and we need not add test cases for every new
events. The existing cases would be sufficient.

### Provide a list of related PRs _(if any)_
NA

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_
`autorest --typescript swagger\README.md`

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)

@xirzec @ellismg @jeremymeng  Please review and approve the PR.

@lmazuel FYI