[ACR] Fix flaky live tests (#25841)
### Packages impacted by this PR

- `@azure/container-registry`

### Describe the problem that is addressed by this PR

Docker manifest upload and download live tests have started failing
recently due to changes in the `library/hello-world` image on Docker
Hub, which we deployed in our test resources. This PR removes the
dependency on the Docker Hub image from our tests so we won't have this
issue any more.