[@azure/eventgrid] Update EG Minor Version (#25897)
### Packages impacted by this PR
@azure/eventgrid

### Issues associated with this PR
None

### Describe the problem that is addressed by this PR
Now new features/bug fixes are added to this PR. The previous release
4.11.1 added new system events to the SDK. Normally, we update the minor
version and not the patch version for such releases. But, since only the
patch version has been updated for that release, in order to maintain
consistency, on the release guidelines, I am updating the minor version
and releasing a new version of the SDK through this PR.

### What are the possible designs available to address the problem? If
there are more than one possible design, why was the one in this PR
chosen?
No special design consideration.

### Are there test cases added in this PR? _(If not, why?)_
Very trivial update of Minor Version. No test cases are required.

### Provide a list of related PRs _(if any)_
None

### Command used to generate this PR:**_(Applicable only to SDK release
request PRs)_
None

### Checklists
- [X] Added impacted package name to the issue description
- [ ] Does this PR needs any fixes in the SDK Generator?** _(If so,
create an Issue in the
[Autorest/typescript](https://github.com/Azure/autorest.typescript)
repository and link it here)_
- [X] Added a changelog (if necessary)

@xirzec Please review and approve. Thanks