[Core] remove URL validations on credential scopes (#25896)
It is possible that credential scopes are not in the format of valid
URL. It is also noted in the doc that there are shorter form of scope
https://learn.microsoft.com/EN-US/azure/active-directory/develop/scopes-oidc

>In requests to the authorization, token or consent endpoints for the
Microsoft Identity platform, if the resource identifier is omitted in
the scope parameter, the resource is assumed to be Microsoft Graph. For
example, `scope=User.Read` is equivalent to
`https://graph.microsoft.com/User.Read`.

Related issue: #25881