[OpenAI] Add library (#25921)
This PR introduces the @azure/ai-openai library. Please ignore the
`sources` folder for now. Also, please note that it has been decided to
use the REST names of the properties in the output models for this first
release, so although they don't sound like the best names, they will not
be renamed in this first release.

Also, please note this package is scheduled for release on 5/22.

---------

Co-authored-by: Jeremy Meng <jeremy.ymeng@gmail.com>