[Test][monitor-query] fix an failing test (#26004)
This particular test sends a query that is supposed to take a long time
and sets a one-second timeout limit, in order to catch the expected
Gateway timeout error. However, it looks that the service returns the
result pretty quickly most of the time, causing the test to fail.

This PR makes the step limit in the query 1000 magintude larger and
randomizes it to avoid getting cached query results.