[Event Hubs] Add stress tests to event-hubs SDK (#25661)
# Stress tests for Event Hubs
- Adds the stress testing framework to the Event Hubs SDK.
- Adds the relevant tests from
https://gist.github.com/HarshaNalluru/2f1931ea30e8335678310053aa342db0
that have helped in the recent memory-leak investigations.
- Follows the patterns from the service-bus stress tests
- In sync with SB after
https://github.com/Azure/azure-sdk-for-js/pull/25658
- tracks message count
- tracks memory metrics

## EventHubsStressTester 
- Sets up Application Insights for logging and tracking events.
- `EventHubsStressTester` class has methods for initializing the test,
taking snapshots of test progress, and ending the test.
- `loopForever` takes a function fn, a delay duration, and an optional
abort signal as arguments. The function enters a loop that calls fn
repeatedly with a delay between each call until the abort signal is
triggered or times out.

## noActivity test
- **Initialization**: The main function `scenarioNoActivity` creates
instances of Event Hubs consumer and producer clients and subscribes to
all partitions of the consumer client with handlers for processing
events, errors, initialization, and closing.
- **Execution**: The producer sends a batch of two events and starts a
loop that checks if the test duration has been reached or if any of the
subscribers have stopped running.
- **Cleanup**: If either condition is met, the loop exits and the
producer and consumer clients are closed. The stress test is ended by
calling the `endTest` method on the `stressBase` instance.

## getRuntimeProperties test
- **Initialization**: The main function `scenarioGetRuntimeProperties`
creates instances of Event Hubs consumer and producer clients and a
`stressBase` instance for stress testing.
- **Execution**: The function defines an asynchronous function `func`
that enters a loop and repeatedly calls the `getEventHubProperties`
method on the consumer client with a random delay between each call. The
function then creates an array of 100 promises that resolve when `func`
completes and waits for all promises to resolve.
- **Cleanup**: Once all promises have resolved, the producer and
consumer clients are closed and the stress test is ended by calling the
`endTest` method on the `stressBase` instance.

## checkpoint store test
- **Initialization**: The main function `scenarioCheckpointStore`
creates instances of Event Hubs consumer and producer clients, a
`stressBase` instance for stress testing, and a `BlobCheckpointStore`
instance for storing checkpoints. The consumer client subscribes to the
default consumer group with handlers for processing events, errors,
initialization, and closing.
- **Execution**: The function enters a loop that sends a batch of random
events using the producer client and waits for a random delay between
each iteration. The loop checks if the test duration has been reached or
if the subscription has stopped running.
- **Cleanup**: If either condition is met, the loop exits and the
producer and consumer clients are closed. The stress test is ended by
calling the `endTest` method on the `stressBase` instance.

---------

Co-authored-by: Deyaaeldeen Almahallawi <dealmaha@microsoft.com>